+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Unreleased
 * Asteroid population blocks now support the log-normal, (rescaled) beta, and gamma distributions.
 * Can now customize asteroids on intercept trajectories.
 * Asteroid .value syntax now supports several characteristic periods and speeds.
+* Asteroids can now appear only under certain conditions.
 * Asteroid orbits can now be given relative to an inclined plane. Useful for mods like RSS and Harder Solar System.
 
 ### Changes